Move Tag generation to a separate class, so we can have separate generators for different classes later on. This also makes DataObjectAnnotator cleaner and easier to rename.
