+++ code/DataObjectAnnotator.php
@@ -1,5 +1,4 @@
 <?php
-use phpDocumentor\Reflection\DocBlock\Tag;
 
 /**
  * Class DataObjectAnnotator
@@ -17,7 +16,6 @@ use phpDocumentor\Reflection\DocBlock\Tag;
  */
 class DataObjectAnnotator extends Object
 {
-
     /**
      * This string marks the beginning of a generated annotations block
      */
@@ -43,21 +41,6 @@ class DataObjectAnnotator extends Object
     private static $enabled_modules = array('mysite');
 
     /**
-     * @var array
-     * Available properties to generate docblocks for.
-     */
-    protected static $propertyTypes = array(
-        'Owner',
-        'DB',
-        'HasOne',
-        'BelongsTo',
-        'HasMany',
-        'ManyMany',
-        'BelongsManyMany',
-        'Extensions',
-    );
-
-    /**
      * @var AnnotatePermissionChecker
      */
     private $permissionChecker;
@@ -69,32 +52,22 @@ class DataObjectAnnotator extends Object
     protected $classes;
 
     /**
-     * All classes that subclass Object
-     * @var array
-     */
-    protected $extensionClasses;
-
-    /**
      * List of all objects, so we can find the extensions.
      * @var array
      */
     protected $dataExtensions;
 
     /**
-     * List all the generated tags form the various generateSomeORMProperies methods
-     * @var array
+     * Temporary flag so we can switch implementations.
+     * Keep it public for easy checking outside this class.
+     * @var bool
      */
-    protected $tags = array(
-        'properties'=> array(),
-        'methods'   => array(),
-        'mixins'    => array()
-    );
+    public static $phpDocumentorEnabled = false;
 
     public function __construct()
     {
         parent::__construct();
         $this->classes = ClassInfo::subclassesFor('DataObject');
-        $this->extensionClasses = ClassInfo::subclassesFor('Object');
         $this->dataExtensions = ClassInfo::subclassesFor('DataExtension');
         $this->permissionChecker = Injector::inst()->get('AnnotatePermissionChecker');
     }
@@ -166,17 +139,12 @@ class DataObjectAnnotator extends Object
      */
     protected function getFileContentWithAnnotations($fileContent, $className)
     {
-        $this->generateORMProperties($className);
+        $generator = new DocBlockTagGenerator($className);
 
-        if (!$this->tags) {
+        $tagString = $generator->getTagsAsString();
+
+        if (!$tagString) {
             return null;
-        }
-
-        $tagString = '';
-        foreach($this->tags as $tagType) {
-            foreach($tagType as $tag) {
-                $tagString .= ' * ' . $tag . "\n";
-            }
         }
 
         $startTag = static::STARTTAG;
@@ -184,17 +152,16 @@ class DataObjectAnnotator extends Object
 
         if (strpos($fileContent, $startTag) && strpos($fileContent, $endTag)) {
             $replacement = $startTag . "\n * \n" . $tagString . " * \n * " . $endTag;
-
             return preg_replace("/$startTag([\s\S]*?)$endTag/", $replacement, $fileContent);
-        } else {
-            $classDeclaration = 'class ' . $className . ' extends'; // add extends to exclude Controller writes
-            $properties = "\n/**\n * " . $startTag . "\n * \n"
-                . $tagString
-                . " * \n * " . $endTag . "\n"
-                . " */\n$classDeclaration";
-
-            return str_replace($classDeclaration, $properties, $fileContent);
         }
+
+        $classDeclaration = 'class ' . $className . ' extends'; // add extends to exclude Controller writes
+        $properties = "\n/**\n * " . $startTag . "\n * \n"
+            . $tagString
+            . " * \n * " . $endTag . "\n"
+            . " */\n$classDeclaration";
+
+        return str_replace($classDeclaration, $properties, $fileContent);
     }
 
     /**
@@ -220,210 +187,5 @@ class DataObjectAnnotator extends Object
     }
 
 
-    /**
-     * @param String $className
-     *
-     * @return string
-     */
-    protected function generateORMProperties($className)
-    {
-        /*
-         * Start with an empty resultstring before generation
-         */
-        $this->resetTags();
 
-        /*
-         * Loop the available types and generate the ORM property.
-         */
-        foreach (self::$propertyTypes as $type) {
-            $function = 'generateORM' . $type . 'Properties';
-            $this->{$function}($className);
-        }
-    }
-
-    /**
-     * Generate the Owner-properties for extensions.
-     *
-     * @param string $className
-     */
-    protected function generateORMOwnerProperties($className)
-    {
-        $owners = array();
-        foreach ($this->extensionClasses as $class) {
-            $config = Config::inst()->get($class, 'extensions', Config::UNINHERITED);
-            if ($config !== null && in_array($className, $config, null)) {
-                $owners[] = $class;
-            }
-        }
-        if (count($owners)) {
-            $owners[] = $className;
-            $tag = implode("|", $owners) . " \$owner";
-            $this->tags['properties'][$tag] = new Tag('property', $tag);
-        }
-    }
-
-    /**
-     * Generate the $db property values.
-     *
-     * @param DataObject|DataExtension $className
-     *
-     * @return string
-     */
-    protected function generateORMDBProperties($className)
-    {
-        if ($fields = Config::inst()->get($className, 'db', Config::UNINHERITED)) {
-            foreach ($fields as $fieldName => $dataObjectName) {
-                $prop = 'string';
-
-                $fieldObj = Object::create_from_string($dataObjectName, $fieldName);
-
-                if ($fieldObj instanceof Int || $fieldObj instanceof DBInt) {
-                    $prop = 'int';
-                } elseif ($fieldObj instanceof Boolean) {
-                    $prop = 'boolean';
-                } elseif ($fieldObj instanceof Float ||
-                    $fieldObj instanceof DBFloat ||
-                    $fieldObj instanceof Decimal
-                ) {
-                    $prop = 'float';
-                }
-                $tag = "$prop \$$fieldName";
-                $this->tags['properties'][$tag] = new Tag('property', $tag);
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Generate the $belongs_to property values.
-     *
-     * @param DataObject|DataExtension $className
-     *
-     * @return string
-     */
-    protected function generateORMBelongsToProperties($className)
-    {
-        if ($fields = Config::inst()->get($className, 'belongs_to', Config::UNINHERITED)) {
-            //$this->resultString .= " * \n";
-            foreach ($fields as $fieldName => $dataObjectName) {
-                $tag = $dataObjectName . " \$$fieldName";
-                $this->tags['methods'][$tag] = new Tag('method', $tag);
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Generate the $has_one property and method values.
-     *
-     * @param DataObject|DataExtension $className
-     *
-     * @return string
-     */
-    protected function generateORMHasOneProperties($className)
-    {
-        if ($fields = Config::inst()->get($className, 'has_one', Config::UNINHERITED)) {
-            foreach ($fields as $fieldName => $dataObjectName) {
-                $tag = "int \${$fieldName}ID";
-                $this->tags['properties'][$tag] = new Tag('property', $tag);
-            }
-            foreach ($fields as $fieldName => $dataObjectName) {
-                $tag = "{$dataObjectName} {$fieldName}()";
-                $this->tags['methods'][$tag] = new Tag('method', $tag);
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Generate the $has_many method values.
-     *
-     * @param DataObject|DataExtension $className
-     *
-     * @return string
-     */
-    protected function generateORMHasManyProperties($className)
-    {
-        if ($fields = Config::inst()->get($className, 'has_many', Config::UNINHERITED)) {
-            //$this->resultString .= " * \n";
-            foreach ($fields as $fieldName => $dataObjectName) {
-                $tag = "DataList|{$dataObjectName}[] {$fieldName}()";
-                $this->tags['methods'][$tag] = new Tag('method', $tag);
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Generate the $many_many method values.
-     *
-     * @param DataObject|DataExtension $className
-     *
-     * @return string
-     */
-    protected function generateORMManyManyProperties($className)
-    {
-        if ($fields = Config::inst()->get($className, 'many_many', Config::UNINHERITED)) {
-            foreach ($fields as $fieldName => $dataObjectName) {
-                $tag = "ManyManyList|{$dataObjectName}[] {$fieldName}()";
-                $this->tags['methods'][$tag] = new Tag('method', $tag);
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Generate the $belongs_many_many method values.
-     *
-     * @param DataObject|DataExtension $className
-     *
-     * @return string
-     */
-    protected function generateORMBelongsManyManyProperties($className)
-    {
-        if ($fields = Config::inst()->get($className, 'belongs_many_many', Config::UNINHERITED)) {
-            foreach ($fields as $fieldName => $dataObjectName) {
-                $tag = "ManyManyList|{$dataObjectName}[] {$fieldName}()";
-                $this->tags['methods'][$tag] = new Tag('method', $tag);
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Generate the mixins for DataExtensions
-     *
-     * @param DataObject|DataExtension $className
-     *
-     * @return string
-     */
-    protected function generateORMExtensionsProperties($className)
-    {
-        if ($fields = Config::inst()->get($className, 'extensions', Config::UNINHERITED)) {
-            //$this->resultString .= " * \n";
-            foreach ($fields as $fieldName) {
-                $this->tags['mixins'][$fieldName] = new Tag('mixin',$fieldName);
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Reset the tag list after each run
-     */
-    protected function resetTags()
-    {
-        $this->tags = array(
-            'properties'=> array(),
-            'methods'   => array(),
-            'mixins'    => array()
-        );
-    }
 }
